icons has been put on wizard form and mingration has been fixed

+++ database/migrations/2022_05_16_000005_create_learning_artifacts_table.php
@@ -23,7 +23,7 @@ return new class extends Migration {
                 'video',
                 'externo',
             ]);
-            $table->decimal('size')->nullable();
+            $table->integer('size')->nullable();
             $table->string('path')->nullable();
             $table->text('description')->nullable();
             $table->boolean('external');
